Extracted LinearisationInterface (temporary) from EKF
